Make SVGs 100% wide
Internet Explorer 11 and below don’t scale SVG images properly when
they’re used in `<img>` tags. See
https://gist.github.com/larrybotha/7881691 for details.

This manifested itself as the SVG images on the product page being
smaller when viewed in IE11 than other browsers.

This commit explicitly sets them the be 100% wide, which seems to fix
the problem.