Make calculation of business hours timezone aware
`replace` doesn’t convert a time from one timezone to another. It just
changes the label that says what timezone a time is in 😬

`.localize` is how we handle these kind of issues in the API (see
https://github.com/alphagov/notifications-api/blob/d0b467b2fb8fab65ee7f483d91b24ea5eb678f4e/app/utils.py#L42-L44 )

So this commit changes the calculation to use `.localize`, and makes the
tests timezone aware to check we’re doing this right.