Fix errant &nbsp; on request to go live page
There’s a good reason for having the `&nbsp;` – it stops GOV.UK Notify
being split across two lines (which could happen on a smaller viewport,
eg mobile). Gotta protect the brand.

Not good for the brand for it to be showing up in the page though 😬

This got broken as part of 3f41090a9428934bef480baceb07b4eae8a84fd4

The label for a form should never have user-submitted content in it, so
using `safe` is fine.