Add a back to template link on job page
When you’ve sent message(s) using a template, often the next thing you
want to do is go and send the same template again, or edit it.

Currently there’s no way of getting to a template from a job except for
going back to the list of templates and re-finding it.

This commit adds a link at the bottom of the job page that gives you a
shortcut back to the individual template, where you can find actions
like edit/send/etc.