Make position of letter edit links absolute
Previously they were relative (ie percentages). This made sure that they
worked on mobile, when the letter might be narrower.

However it broke when the preview was more than one page, because
13% of the height of 2 pages is different to 13% of the height of one
pages.

This commit changes the positions to be pixel values, which match the
calculated percentage values when the preview is one page.