Show spreadsheet before template if there’s errors
It’s easier to work out what the errors are in your file if you can
see the big red error message right next to the contents of your file.

When users get errors they tend to go back and forth between the error
message and the view of the file, not the message and the template
preview.