Use black text with a border for default banner
We use panels with a blue banner to indicate something that’s clickable.
So we should move away from this style for things that are just
notifications. We can’t use teal like other bits of GOV.UK because it
doesn’t pass colour contrast.

Pay are using a box with a green border, similar to the error validation
box (which has a red border). So let’s do the same for now.